Modularize & deprecate rates
