+++ openfisca_core/commons/__init__.py
@@ -23,10 +23,6 @@
 
 from .dummy import Dummy  # noqa: F401
 
-from .helpers import (  # noqa: F401
-    apply_thresholds,
-    concat,
-    empty_clone,
-    stringify_array,
-    switch,
-    )
+from .formulas import apply_thresholds, concat, switch  # noqa: F401
+from .misc import empty_clone, stringify_array  # noqa: F401
+from .rates import average_rate, marginal_rate  # noqa: F401
